Extract `CatsEffectAssertions` to its own file
