+++ core/shared/src/main/scala/munit/CatsEffectSuite.scala
@@ -20,31 +20,7 @@ import cats.effect.{ContextShift, IO, SyncIO, Timer}
 
 import scala.concurrent.Future
 
-abstract class CatsEffectSuite extends FunSuite {
-
-  /**
-    * Asserts that an [[IO]] returns an expected value.
-    *
-    * The "returns" value (second argument) must have the same type or be a
-    * subtype of the one "contained" inside the `IO` (first argument). For example:
-    * {{{
-    *   assertIO(IO(Option(1)), returns = Some(1)) // OK
-    *   assertIO(IO(Some(1)), returns = Option(1)) // Error: Option[Int] is not a subtype of Some[Int]
-    * }}}
-    *
-    * The "clue" value can be used to give extra information about the failure in case the
-    * assertion fails.
-    *
-    * @param obtained the IO under testing
-    * @param returns the expected value
-    * @param clue a value that will be printed in case the assertions fails
-    */
-  def assertIO[A, B](
-      obtained: IO[A],
-      returns: B,
-      clue: => Any = "values are not the same"
-  )(implicit loc: Location, ev: B <:< A): IO[Unit] =
-    obtained.flatMap(a => IO(assertEquals(a, returns, clue)))
+abstract class CatsEffectSuite extends FunSuite with CatsEffectAssertions {
 
   implicit def munitContextShift: ContextShift[IO] =
     IO.contextShift(munitExecutionContext)
